Refactor server info screen
