+++ lib/functions/format.dart
@@ -21,3 +21,10 @@ String formatTimestampForChart(String timestamp) {
   final end = time.add(const Duration(minutes: 5));
   return '${formatNumber(start.hour)}:${formatNumber(start.minute)} - ${formatNumber(end.hour)}:${formatNumber(end.minute)}';
 }
+
+String formatUptime(int seconds) {
+  final days = seconds ~/ (24 * 3600);
+  final hours = (seconds % (24 * 3600)) ~/ 3600;
+  final minutes = (seconds % 3600) ~/ 60;
+  return '$days d $hours h $minutes m';
+}
