+++ lib/screens/settings/server_settings/server_info.dart
@@ -1,315 +1,20 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_svg/flutter_svg.dart';
 import 'package:pi_hole_client/classes/custom_scroll_behavior.dart';
-import 'package:pi_hole_client/config/theme.dart';
 import 'package:pi_hole_client/functions/logger.dart';
-import 'package:pi_hole_client/functions/misc.dart';
 import 'package:pi_hole_client/gateways/api_gateway_interface.dart';
 import 'package:pi_hole_client/l10n/generated/app_localizations.dart';
-import 'package:pi_hole_client/models/host.dart';
-import 'package:pi_hole_client/models/sensors.dart';
 import 'package:pi_hole_client/models/server.dart';
-import 'package:pi_hole_client/models/system.dart';
 import 'package:pi_hole_client/providers/servers_provider.dart';
-import 'package:pi_hole_client/widgets/adaptive_trailing_text.dart';
+import 'package:pi_hole_client/screens/settings/server_settings/widgets/server_info/host_information_section.dart';
+import 'package:pi_hole_client/screens/settings/server_settings/widgets/server_info/performance_usage_section.dart';
+import 'package:pi_hole_client/screens/settings/server_settings/widgets/server_info/pihole_version_section.dart';
+import 'package:pi_hole_client/screens/settings/server_settings/widgets/server_info/server_connection_section.dart';
 import 'package:pi_hole_client/widgets/empty_data_screen.dart';
 import 'package:pi_hole_client/widgets/error_message_screen.dart';
-import 'package:pi_hole_client/widgets/list_tile_title.dart';
-import 'package:pi_hole_client/widgets/section_label.dart';
 import 'package:provider/provider.dart';
 
 class ServerInfoScreen extends StatelessWidget {
   const ServerInfoScreen({super.key});
-
-  Widget piholeIcon(BuildContext context) {
-    return SvgPicture.asset(
-      'assets/resources/pihole.svg',
-      colorFilter: ColorFilter.mode(
-        Theme.of(context).colorScheme.onSurfaceVariant,
-        BlendMode.srcIn,
-      ),
-      width: 24,
-      height: 24,
-    );
-  }
-
-  Widget basicInfoSection(
-    BuildContext context,
-    Server? server,
-  ) {
-    final colorScheme = Theme.of(context).colorScheme;
-
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        SectionLabel(label: AppLocalizations.of(context)!.connectingServer),
-        ListTile(
-          dense: true,
-          leading: const Icon(Icons.storage_outlined),
-          title: listTileTitle(
-            server?.alias ?? AppLocalizations.of(context)!.unknown,
-            colorScheme: colorScheme,
-          ),
-          subtitle: Text(
-            server?.address ?? AppLocalizations.of(context)!.unknown,
-            maxLines: 1,
-            overflow: TextOverflow.ellipsis,
-          ),
-          trailing: Icon(
-            Icons.check_circle_rounded,
-            color: Theme.of(context).extension<AppColors>()!.commonGreen,
-          ),
-        ),
-      ],
-    );
-  }
-
-  Widget hostInfoSection(
-    BuildContext context,
-    HostInfo? host,
-  ) {
-    final colorScheme = Theme.of(context).colorScheme;
-
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        SectionLabel(label: AppLocalizations.of(context)!.host),
-        ListTile(
-          dense: true,
-          leading: const Icon(Icons.connected_tv_rounded),
-          title: listTileTitle(
-            AppLocalizations.of(context)!.host,
-            colorScheme: colorScheme,
-          ),
-          trailing: AdaptiveTrailingText(
-            text: host?.hostName ?? AppLocalizations.of(context)!.unknown,
-          ),
-        ),
-        ListTile(
-          dense: true,
-          leading: const Icon(Icons.tag_rounded),
-          title: listTileTitle(
-            AppLocalizations.of(context)!.model,
-            colorScheme: colorScheme,
-          ),
-          trailing: AdaptiveTrailingText(
-            text: host?.model ?? AppLocalizations.of(context)!.unknown,
-          ),
-        ),
-        Theme(
-          data: Theme.of(context).copyWith(dividerColor: Colors.transparent),
-          child: ExpansionTile(
-            title: listTileTitle(
-              AppLocalizations.of(context)!.moreDetails,
-              colorScheme: colorScheme,
-            ),
-            children: [
-              ListTile(
-                dense: true,
-                leading: const Icon(Icons.settings_rounded),
-                title: listTileTitle(
-                  AppLocalizations.of(context)!.system,
-                  colorScheme: colorScheme,
-                ),
-                trailing: AdaptiveTrailingText(
-                  text: host?.sysName ?? AppLocalizations.of(context)!.unknown,
-                ),
-              ),
-              ListTile(
-                dense: true,
-                leading: const Icon(Icons.architecture_rounded),
-                title: listTileTitle(
-                  AppLocalizations.of(context)!.arch,
-                  colorScheme: colorScheme,
-                ),
-                trailing: AdaptiveTrailingText(
-                  text: host?.arch ?? AppLocalizations.of(context)!.unknown,
-                ),
-              ),
-              ListTile(
-                dense: true,
-                leading: const Icon(Icons.system_update_rounded),
-                title: listTileTitle(
-                  AppLocalizations.of(context)!.release,
-                  colorScheme: colorScheme,
-                ),
-                trailing: AdaptiveTrailingText(
-                  text: host?.release ?? AppLocalizations.of(context)!.unknown,
-                ),
-              ),
-              ListTile(
-                dense: true,
-                leading: const Icon(Icons.verified_rounded),
-                title: listTileTitle(
-                  AppLocalizations.of(context)!.version,
-                  colorScheme: colorScheme,
-                ),
-                trailing: AdaptiveTrailingText(
-                  text: host?.version ?? AppLocalizations.of(context)!.unknown,
-                ),
-              ),
-            ],
-          ),
-        ),
-      ],
-    );
-  }
-
-  Widget usageSection(
-    BuildContext context,
-    SystemInfo? system,
-    SensorsInfo? sensors,
-  ) {
-    final colorScheme = Theme.of(context).colorScheme;
-
-    String formatUptime(int seconds) {
-      final days = seconds ~/ (24 * 3600);
-      final hours = (seconds % (24 * 3600)) ~/ 3600;
-      final minutes = (seconds % 3600) ~/ 60;
-      return '$days d $hours h $minutes m';
-    }
-
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        SectionLabel(label: AppLocalizations.of(context)!.performance),
-        ListTile(
-          dense: true,
-          leading: const Icon(Icons.memory_rounded),
-          title: listTileTitle(
-            AppLocalizations.of(context)!.cpuUsage,
-            colorScheme: colorScheme,
-          ),
-          subtitle: Column(
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: [
-              LinearProgressIndicator(
-                value: (system?.cpuUsage ?? 0.0) / 100.0,
-                valueColor: AlwaysStoppedAnimation<Color>(
-                  Theme.of(context).colorScheme.primary,
-                ),
-                backgroundColor: Theme.of(context).colorScheme.primaryContainer,
-              ),
-              const SizedBox(height: 4),
-              Text(
-                '${(system?.cpuUsage ?? 0.0).toStringAsFixed(2)}%',
-              ),
-            ],
-          ),
-        ),
-        ListTile(
-          dense: true,
-          leading: const Icon(Icons.speed_rounded),
-          title: listTileTitle(
-            AppLocalizations.of(context)!.memoryUsage,
-            colorScheme: colorScheme,
-          ),
-          subtitle: Column(
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: [
-              LinearProgressIndicator(
-                value: (system?.ramUsage ?? 0.0) / 100.0,
-                valueColor: AlwaysStoppedAnimation<Color>(
-                  Theme.of(context).colorScheme.primary,
-                ),
-                backgroundColor: Theme.of(context).colorScheme.primaryContainer,
-              ),
-              const SizedBox(height: 4),
-              Text(
-                '${(system?.ramUsage ?? 0.0).toStringAsFixed(2)}%',
-              ),
-            ],
-          ),
-        ),
-        Theme(
-          data: Theme.of(context).copyWith(dividerColor: Colors.transparent),
-          child: ExpansionTile(
-            title: listTileTitle(
-              AppLocalizations.of(context)!.moreDetails,
-              colorScheme: colorScheme,
-            ),
-            children: [
-              ListTile(
-                dense: true,
-                leading: const Icon(Icons.access_time),
-                title: listTileTitle(
-                  AppLocalizations.of(context)!.uptime,
-                  colorScheme: colorScheme,
-                ),
-                trailing: AdaptiveTrailingText(
-                  text: system != null
-                      ? formatUptime(system.uptime)
-                      : AppLocalizations.of(context)!.unknown,
-                ),
-              ),
-              ListTile(
-                dense: true,
-                leading: const Icon(Icons.thermostat_rounded),
-                title: listTileTitle(
-                  AppLocalizations.of(context)!.cpuTemperature,
-                  colorScheme: colorScheme,
-                ),
-                trailing: AdaptiveTrailingText(
-                  text: sensors?.cpuTemp != null
-                      ? '${(sensors?.cpuTemp ?? 0.0).toStringAsFixed(2)} ${convertTemperatureUnit(sensors?.unit)}'
-                      : AppLocalizations.of(context)!.unknown,
-                ),
-              ),
-            ],
-          ),
-        ),
-      ],
-    );
-  }
-
-  Widget piholeVersionSection(
-    BuildContext context,
-    dynamic version,
-  ) {
-    final colorScheme = Theme.of(context).colorScheme;
-
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        SectionLabel(label: AppLocalizations.of(context)!.piholeVersion),
-        ListTile(
-          dense: true,
-          leading: piholeIcon(context),
-          title: listTileTitleNoPadding('Core', colorScheme: colorScheme),
-          subtitle: Text('${version?.core.local.branch ?? '-'}'),
-          trailing:
-              AdaptiveTrailingText(text: version?.core.local.version ?? '-'),
-        ),
-        ListTile(
-          dense: true,
-          leading: piholeIcon(context),
-          title: listTileTitleNoPadding('FTL', colorScheme: colorScheme),
-          subtitle: Text('${version?.ftl.local.branch ?? '-'}'),
-          trailing:
-              AdaptiveTrailingText(text: version?.ftl.local.version ?? '-'),
-        ),
-        ListTile(
-          dense: true,
-          leading: piholeIcon(context),
-          title: listTileTitleNoPadding('Web', colorScheme: colorScheme),
-          subtitle: Text('${version?.web.local.branch ?? '-'}'),
-          trailing:
-              AdaptiveTrailingText(text: version?.web.local.version ?? '-'),
-        ),
-        ListTile(
-          dense: true,
-          leading: piholeIcon(context),
-          title: listTileTitleNoPadding('Docker', colorScheme: colorScheme),
-          subtitle: const Text('-'),
-          trailing: AdaptiveTrailingText(
-            text: (version?.docker.local?.isNotEmpty ?? false)
-                ? version!.docker.local!
-                : '-',
-          ),
-        ),
-      ],
-    );
-  }
 
   @override
   Widget build(BuildContext context) {
@@ -355,16 +60,16 @@ class ServerInfoScreen extends StatelessWidget {
               child: Column(
                 crossAxisAlignment: CrossAxisAlignment.start,
                 children: [
-                  basicInfoSection(context, server),
+                  ServerConnectionSection(server: server),
                   const SizedBox(height: 20),
                   // v5 API not support host info and usage info
                   if (apiGateway?.server.apiVersion != 'v5') ...[
-                    hostInfoSection(context, host),
+                    HostInformationSection(host: host),
                     const SizedBox(height: 20),
-                    usageSection(context, system, sensor),
+                    PerformanceUsageSection(system: system, sensors: sensor),
                     const SizedBox(height: 20),
                   ],
-                  piholeVersionSection(context, version),
+                  PiholeVersionSection(version: version),
                 ],
               ),
             );
