+++ lib/screens/settings/server_settings/widgets/server_info/performance_usage_section.dart
@@ -1,0 +1,113 @@
+import 'package:flutter/material.dart';
+import 'package:pi_hole_client/functions/format.dart';
+import 'package:pi_hole_client/functions/misc.dart';
+import 'package:pi_hole_client/l10n/generated/app_localizations.dart';
+import 'package:pi_hole_client/models/sensors.dart';
+import 'package:pi_hole_client/models/system.dart';
+import 'package:pi_hole_client/widgets/adaptive_trailing_text.dart';
+import 'package:pi_hole_client/widgets/list_tile_title.dart';
+import 'package:pi_hole_client/widgets/section_label.dart';
+
+class PerformanceUsageSection extends StatelessWidget {
+  const PerformanceUsageSection({
+    required this.system,
+    required this.sensors,
+    super.key,
+  });
+
+  final SystemInfo? system;
+  final SensorsInfo? sensors;
+
+  @override
+  Widget build(BuildContext context) {
+    final colorScheme = Theme.of(context).colorScheme;
+
+    return Column(
+      crossAxisAlignment: CrossAxisAlignment.start,
+      children: [
+        SectionLabel(label: AppLocalizations.of(context)!.performance),
+        ListTile(
+          dense: true,
+          leading: const Icon(Icons.memory_rounded),
+          title: listTileTitle(
+            AppLocalizations.of(context)!.cpuUsage,
+            colorScheme: colorScheme,
+          ),
+          subtitle: Column(
+            crossAxisAlignment: CrossAxisAlignment.start,
+            children: [
+              LinearProgressIndicator(
+                value: (system?.cpuUsage ?? 0.0) / 100.0,
+                valueColor: AlwaysStoppedAnimation<Color>(
+                  Theme.of(context).colorScheme.primary,
+                ),
+                backgroundColor: Theme.of(context).colorScheme.primaryContainer,
+              ),
+              const SizedBox(height: 4),
+              Text('${(system?.cpuUsage ?? 0.0).toStringAsFixed(2)}%'),
+            ],
+          ),
+        ),
+        ListTile(
+          dense: true,
+          leading: const Icon(Icons.speed_rounded),
+          title: listTileTitle(
+            AppLocalizations.of(context)!.memoryUsage,
+            colorScheme: colorScheme,
+          ),
+          subtitle: Column(
+            crossAxisAlignment: CrossAxisAlignment.start,
+            children: [
+              LinearProgressIndicator(
+                value: (system?.ramUsage ?? 0.0) / 100.0,
+                valueColor: AlwaysStoppedAnimation<Color>(
+                  Theme.of(context).colorScheme.primary,
+                ),
+                backgroundColor: Theme.of(context).colorScheme.primaryContainer,
+              ),
+              const SizedBox(height: 4),
+              Text('${(system?.ramUsage ?? 0.0).toStringAsFixed(2)}%'),
+            ],
+          ),
+        ),
+        Theme(
+          data: Theme.of(context).copyWith(dividerColor: Colors.transparent),
+          child: ExpansionTile(
+            title: listTileTitle(
+              AppLocalizations.of(context)!.moreDetails,
+              colorScheme: colorScheme,
+            ),
+            children: [
+              ListTile(
+                dense: true,
+                leading: const Icon(Icons.access_time),
+                title: listTileTitle(
+                  AppLocalizations.of(context)!.uptime,
+                  colorScheme: colorScheme,
+                ),
+                trailing: AdaptiveTrailingText(
+                  text: system != null
+                      ? formatUptime(system!.uptime)
+                      : AppLocalizations.of(context)!.unknown,
+                ),
+              ),
+              ListTile(
+                dense: true,
+                leading: const Icon(Icons.thermostat_rounded),
+                title: listTileTitle(
+                  AppLocalizations.of(context)!.cpuTemperature,
+                  colorScheme: colorScheme,
+                ),
+                trailing: AdaptiveTrailingText(
+                  text: sensors?.cpuTemp != null
+                      ? '${(sensors?.cpuTemp ?? 0.0).toStringAsFixed(2)} ${convertTemperatureUnit(sensors?.unit)}'
+                      : AppLocalizations.of(context)!.unknown,
+                ),
+              ),
+            ],
+          ),
+        ),
+      ],
+    );
+  }
+}
